+++ api/database-conn.php
@@ -1,8 +1,8 @@
-<!-- This is a file that creates a database connection, used by all other files. -->
-<!--
+<?php /* 
+This is a file that creates a database connection, used by all other files.
+
 Copyright Antoni Tyczka 2024
--->
-<?php
+*/
 if(__FILE__ == $_SERVER["SCRIPT_FILENAME"]){
 	//called directly
 	http_response_code(405);